[BACKLOG-19904] Isolated logic to get resourceBundles inside a new service
