+++ pentaho-i18n-bundle/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -3,7 +3,24 @@
            xsi:schemaLocation="
   http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd">
 
-  <bean id="localizationManager" class="org.pentaho.osgi.i18n.impl.LocalizationManager" scope="singleton"/>
-  <service ref="localizationManager" interface="org.pentaho.osgi.i18n.LocalizationService"/>
+  <!-- LocalizationManager -->
+  <bean id="localizationServiceImpl" class="org.pentaho.osgi.i18n.impl.LocalizationManager" scope="singleton"/>
+  <service ref="localizationServiceImpl"
+           interface="org.pentaho.osgi.i18n.LocalizationService"/>
+
+
+  <!-- PentahoWebPackageLocalizationService -->
+  <reference id="pentahoWebPackageService" interface="org.pentaho.webpackage.core.PentahoWebPackageService"/>
+
+  <bean id="pentahoWebPackageLocalizationService"
+        class="org.pentaho.osgi.i18n.impl.PentahoWebPackageLocalizationService"
+        scope="singleton">
+    <property name="localizationService" ref="localizationServiceImpl"/>
+    <property name="webPackageService" ref="pentahoWebPackageService"/>
+  </bean>
+
+  <service ref="pentahoWebPackageLocalizationService"
+           interface="org.pentaho.osgi.i18n.IPentahoWebPackageLocalizationService"/>
+
 
 </blueprint>
