+++ pentaho-i18n-webservice-bundle/src/main/java/org/pentaho/osgi/i18n/webservice/ResourceBundleMessageBodyWriter.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.osgi.i18n.webservice.impl;
+package org.pentaho.osgi.i18n.webservice;
 
 import org.json.simple.JSONObject;
 import org.w3c.dom.Document;
